Added type validations

+++ src/actions/dist.ts
@@ -4,7 +4,7 @@ import { exec } from 'child_process';
 import fs from 'fs';
 
 export function dist(callback: () => void, watch = false) {
-  if (fs.existsSync(tsconfig.compilerOptions.outDir)) {
+  if (tsconfig.compilerOptions.outDir && fs.existsSync(tsconfig.compilerOptions.outDir)) {
     fs.rmSync(tsconfig.compilerOptions.outDir, { recursive: true });
   }
   exec(`tsc ${watch ? '-w' : ''} --project ${config.esbuildOptions.tsconfig}`, callback);
