+++ src/utils/setupBuild.ts
@@ -5,11 +5,14 @@ import { injectServiceWorker } from '../utils/injectServiceWorker';
 import { minifyHTML } from '../utils/minifyHTML';
 
 export const setupBuild = () => {
-  if (fs.existsSync(config.esbuildOptions.outdir)) {
-    fs.rmSync(config.esbuildOptions.outdir, { recursive: true });
+  if (config.esbuildOptions?.outdir) {
+    if (fs.existsSync(config.esbuildOptions.outdir)) {
+      fs.rmSync(config.esbuildOptions.outdir, { recursive: true });
+    }
+    fs.mkdirSync(config.esbuildOptions.outdir, { recursive: true })
+    const indexTranformer = (fileContent: string) => config.public.minifyIndex ? minifyHTML(fileContent) : fileContent;
+    if (config.public.indexName)
+      copy(config.public.path, config.esbuildOptions.outdir, [{ fileName: config.public.indexName, transformer: indexTranformer }]);
+    injectServiceWorker();
   }
-  fs.mkdirSync(config.esbuildOptions.outdir, {recursive: true})
-  const indexTranformer = (fileContent: string) => config.public.minifyIndex ? minifyHTML(fileContent) : fileContent;
-  copy(config.public.path, config.esbuildOptions.outdir, [{ fileName: config.public.indexName, transformer: indexTranformer }]);
-  injectServiceWorker();
 }
